chore: ELv2 license key functionality cleanup
This commit does not change the licensing of this project, but it does aim to simplify and bring clarity and maintain a free and open core to the Router.  This simplification is offered via a couple changes to so-called "license key functionality", which is the mechanism which enables "GraphOS Enterprise" features in the Router for Enterprise customers.

- Changes terminology within the codebase from "entitlement" to "license" to
  avoid unnecessary indirection between the terms.
- Remove superfluous "license key functionality" annotations throughout the
  codebase.