Query planner plugins (#3177)
Fixes #3150
Related: https://github.com/apollographql/apollo-rs/issues/420

This reintroduces the query planner plugins we had in the past, but with
a different API, and for internal use only for now, until we are sure
about the API's shape.

## Plugin API

The plugins need access to a compiler instance to observe and modify the
query, so the caching query planner adds it as part of the query planner
request (which means that now the cache and bridge get different request
types). For now, if a plugin needs to modify the query, it must
serialize the modified version to a string then reparse it, but the goal
is to support modification right inside the query:
https://github.com/apollographql/apollo-rs/issues/420
That compiler will hold both the schema's type info, and the query.

Once the requests reaches the bridge query planner, it will generate:
- a query plan from the filtered query
- selections from the filtered query (the `Query` object used for
response formatting)
- selections from the original query

Thus execution will depend on the filtered query, so if some fields are
added or removed, it wil change the generated query plan. The plan and
selections will then be cached in the same way as before.

## Details on selections

Query planner plugins can modify the query before sending it through
query planning. They might require different sets of fields on the same
entity, and the query plan might add its own list of fields (keys for
federated queries).
As an example, We could have a User entity with its email field used as
key, and want to remove private information from the query.

So we could have the query:

```graphql
{
  topProducts {
    name
    reviews {
      author {
        email
        username
      }
    }
  }
}
```

So here we would filter the query as follows:

```graphql
{
  topProducts {
    name
    reviews {
      author {
        username
      }
    }
  }
}
```

But since the email is used as key, it would appear in the JSON object
that accumulates response data.
To avoid sending non requested data to the client, we have the response
formatting phase, that uses selections extracted from the query, to pick
only what is needed from the JSON object, and send it to the client.

Here we need to apply the selections of the filtered query, to make sure
the email is not returned to the client (the original query would let it
go through). But we also need to apply the selections from the original
query, to have the null propagation algorithm apply and return a
response in the shape that matches the client query.

There is probably a better way to do it than applying selections twice,
but here we are sure the behaviour will be correct, and so far the
formatting phase is fast enough, we can spend a bit more time there

Co-authored-by: Simon Sapin <simon@apollographql.com>
Co-authored-by: Jeremy Lempereur <jeremy.lempereur@iomentum.com>
Co-authored-by: Gary Pennington <gary@apollographql.com>