Improve Rhai errors (#3254)
This PR attempts to improve the Rhai error handling and debugging
experience. The only tweak to the code is that if there is an error in
the main script, the path to that script is included in the error.
Previously, the plugin name `apollo.rhai` (which _looks_ like a file
name for a script) combined with the fact that line numbers/positions
are incorrect in syntax errors, made it seem like some internal
generated code was failing.

This does _not_ add paths to errors in imported modules, since the
module names seemed clear enough to me.

Additionally, I added a "Debugging" section to the Rhai docs—this needs
some additions from folks who know Rhai better than I do, but it has
some basics to get started. Specifically, I think we should add some
advice on testing Rhai scripts.

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [x] Changes are compatible[^1]
- [x] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [x] Unit Tests
    - [x] Integration Tests
    - [x] Manual Tests

---------

Co-authored-by: Dylan Anthony <dbanty@users.noreply.github.com>
Co-authored-by: Gary Pennington <garypen@gmail.com>