try to come up with a compromise for k8s
Our health check returns a 200 with a JSON payload. It isn't easy to
process that in k8s. Especially when you can't install useful utilities
into your container image (like curl, jq, etc...) because of security
and CVEs.

Compromise: when the router is not live or ready, return a 503 which k8s
interprets as a probe fail.