add additional build functionality to the diy build script (#3304)
The diy build script is useful for ad-hoc image creation during testing
or for building your own images based on a router repo. This set of
enhancements makes it possible to

 - build docker images from arbitrary (nightly) builds (-a)
 - build an amd64 docker image on an arm64 machine (or vice versa) (-m)
 - change the name of the image from the default 'router' (-n)

Note: the build machine architecture is used if the -m flag is not
supplied.

fixes: #3303

<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [x] Changes are compatible[^1]
- [x] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`