add labels to deferred fragments when they are missing (#3298)
We currently use a path and subselection to identify deferred fragments
and format individual responses. Unfortunately, with query filtering,
matching those fragments between the original and filtered query is hard
because they can have a different shape, and thus a different
subselection. Additionally, we need different response shape variants
depending on query variables, due to the `@defer` directive's `if`
argument.
As an alternative solution, we can add a label to every deferred
fragment. The label argument in the `@defer` directive is optional, but
unique across the entire query. Some clients use it for that exact
purpose, identifying each deferred response.
So the idea here is to add labels when they are not present, at the
query plan cache level, use all the labels to identify the fragments, and
then in the execution service, remove the extraneous labels from the
responses sent to the client

Co-authored-by: Simon Sapin <simon@apollographql.com>
Co-authored-by: o0Ignition0o <jeremy.lempereur@gmail.com>