Update example for claim forwarding (#3319)
The JWT claim example that we had in our docs was insecure as it
iterated over the list of claims and set them as headers.
A malicious user could have provided a valid JWT that was missing claims
and then set those claims as headers.
This would only have affected users who had configured their routers to
forward all headers from the client to subgraphs.

The documentation has been updated to explicitly list the claims that
are forwarded to the subgraph.
In addition, a new example has been added that uses extensions to
forward claims.
Fixes #3224

<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [ ] Changes are compatible[^1]
- [x] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [x] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`

Co-authored-by: bryn <bryn@apollographql.com>