add expansion module support to rhai (#3240)
This refactors existing functionality within `Expansion` to expose a new
function for environment variable access. This leverages the existing
mechanism for limiting access to environment variables.

This functionality is then exposed to rhai scripts as in this example
code:

```
    try {
        print(`LANG: ${env::get("LANG")}`);
    } catch(err) {
        print(`exception: ${err}`);
    }
```

In the sample, we use the new env module to access environment
variables, in this case "LANG".

fixes: #1744

<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [x] Changes are compatible[^1]
- [x] Documentation[^2] completed
- [x] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [x] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`