fix response formatting with defer and query filtering
With a query contains a deferred fragment, and that deferred fragment
containing a non nullable field, like this one:

```graphql
{
  a {
    b
    ... @defer {
      c
    }
  }
}
```

If the filter removes that non nullable `c` field, then the filtered
query has this shape:

```graphql
{
  a {
    b
  }
}
```

Which will not be deferred, so only one response will be sent, without
multipart. This is fine, because the server can decide if something must
be deferred or not.

But this was triggering issues in response formatting, because response
formatting for the original query was working from the original query,
with the fragment, but ignoring the `@defer` directive.

Which means that if the response ends up being like this after the
filtered query's formatting:

```json
{
  "a": {
    "b": 1,
  }
}
```

It was then formatted according to this shape:

```graphql
{
  a {
    b
    ... {
      c
    }
  }
}
```

which means that ullifying `c`c would propagate to nullifying `a` and
result in:

```json
{
  "a": null
}
```

To fix this, we look at the deferred status, not from what the query
planner reports, but from what each of the original or filtered query
would consider deferred, depending on the presence of fragments with
`@defer`, and the activated variables