Enable serde_json float_roundtrip feature (#3338)
This feature is explicitly designed to:

"""
Use sufficient precision when parsing fixed precision floats from JSON
to ensure that they maintain accuracy when round-tripped through JSON.
This comes at an approximately 2x performance cost for parsing floats
compared to the default best-effort precision.
"""

Despite the performance impact for floats, we need the fix in order to
prevent float values losing precision when processed by the router.

fixes: #2951

<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [x] Changes are compatible[^1]
- [x] Documentation[^2] completed
- [x] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [x] Manual Tests

**Exceptions**

The manual testing was performed as described in the issue: #2951 

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`