Improved readiness/liveness checks (#3276)
The router is a complex beast and it's not entirely straightforward to
determine if a router is live/ready.

From a kubernetes point of view, a service is:

- live [if it isn't
deadlocked](https://www.linkedin.com/posts/llarsson_betterdevopsonkubernetes-devops-devsecops-activity-7018587202121076736-LRxE)
 - ready if it is able to start accepting traffic

(For more details:
https://kubernetes.io/docs/tasks/configure-pod-container/configure-liveness-readiness-startup-probes/)

Our current health check isn't actually doing either of these things.
It's returning a payload which indicates if the router is "healthy" or
not and it's always returning "UP" (hard-code) with HTTP status code:
200. As long as it is serving traffic it appears "healthy".

To address this we need:

1. A mechanism within the router to define whether it is ready or live.
2. A new interface whereby an interested party may determine if the
router is ready or live.

Both of these questions are significantly complicated by the fact that
the router is architected to maintain connectivity in the face of
changing configuration, schemas, etc... Multiple interested parties may
be operating against the same router with different states (via
different connections) and thus could, in theory, get different answers.
However, I'm working on the assumption that, for k8s control purposes, a
client, would only be interested in the current (latest) state of the
router, so that's what we'll track.

---

1. How do we know if the router is ready or live?

Ready

 - Not currently processing a change to its set of input
 - Is in state Running
- Resources available (open files limit not reached) (although this
removes a pod from service, which would be bad)
 - Is live

Live
 - Health check enabled and responding

2. Extending our existing health interface

If we add support for query parameters named "ready" and "live" to our
existing health endpoint, then we can provide a backwards compatible
interface which leverages the existing "/health" endpoint. We would
support both POST and GET for maximum compatibility.

Sample queries:

```
curl -XPOST "http://localhost:8088/health?ready" OR curl  "http://localhost:8088/health?ready"

curl -XPOST "http://localhost:8088/health?live" OR curl "http://localhost:8088/health?live"
```

You could only query for one at a time (which I think is fine).

Note: Our health check returns a 200 with a JSON payload. It isn't easy
to process that in k8s. Especially when you can't install useful
utilities into your container image (like curl, jq, etc...) because of
security and CVEs. When the router is not live or ready, return a 503
which k8s interprets as a probe fail.

Fixes #3233 

<!-- start metadata -->

**Checklist**

Complete the checklist (and note appropriate exceptions) before a final
PR is raised.

- [x] Changes are compatible[^1]
- [ ] Documentation[^2] completed
- [ ] Performance impact assessed and acceptable
- Tests added and passing[^3]
    - [ ] Unit Tests
    - [ ] Integration Tests
    - [ ] Manual Tests

**Exceptions**

*Note any exceptions here*

**Notes**

[^1]. It may be appropriate to bring upcoming changes to the attention
of other (impacted) groups. Please endeavour to do this before seeking
PR approval. The mechanism for doing this will vary considerably, so use
your judgement as to how and when to do this.
[^2]. Configuration is an important part of many changes. Where
applicable please try to document configuration examples.
[^3]. Tick whichever testing boxes are applicable. If you are adding
Manual Tests:
- please document the manual testing (extensively) in the Exceptions.
- please raise a separate issue to automate the test and label it (or
ask for it to be labeled) as `manual test`

---------

Co-authored-by: Bryn Cooke <BrynCooke@gmail.com>
Co-authored-by: Pierre Tondereau <pierre.tondereau@protonmail.com>
Co-authored-by: Avery Harnish <avery@apollographql.com>
Co-authored-by: bryn <bryn@apollographql.com>
Co-authored-by: renovate[bot] <29139614+renovate[bot]@users.noreply.github.com>
Co-authored-by: Geoffroy Couprie <apollo@geoffroycouprie.com>
Co-authored-by: Simon Sapin <simon@apollographql.com>
Co-authored-by: o0Ignition0o <jeremy.lempereur@gmail.com>
Co-authored-by: Jesse Rosenberger <git@jro.cc>
Co-authored-by: Jeremy Lempereur <jeremy.lempereur@iomentum.com>
Co-authored-by: Dylan Anthony <43723790+dbanty@users.noreply.github.com>