implement the policy directive
Definition:

```graphql
directive @policy(policy: String!) on OBJECT | FIELD_DEFINITION | INTERFACE | SCALAR | ENUM
```

`@policy` is designed for usage with coprocessors:
- extract the list of policies relevant to the query, store them in the
  context
- the coprocessor (or a rhai or native plugin) goes through the list of
  policies and marks them as successful or not
- the router then filters fields from the query according to which
  policies were successful

the `policy` argument could be the actual authorization policy to
execute, in text form, or an index into a list of policies that the
coprocessor knows how to execute.
This will allow router authorization to leverage existing authorization
systems, with custom policy languages, or ones that call into central
state like a roles database.

Field filtering and null propagation happens in exactly the same way as
the other authorization directives, and can be used with them in the
same schema