fix existing bug in lifecyle test
The IngtegrationTest object should be created fresh for each loop so
that the log is fresh for every loop iteration.

The bug wasn't detected previously because of the different log
messages it searched for, but the change tickled the bug because it
found the messages in the wrong place in the log.